INFUND-2135 initial version of view for comp execs to see application summaries

+++ ifs-data-service/src/main/java/com/worth/ifs/application/mapper/ApplicationSummaryMapper.java
@@ -22,7 +22,7 @@ public abstract class ApplicationSummaryMapper {
 		
 		ServiceResult<CompletedPercentageResource> percentageResult = applicationService.getProgressPercentageByApplicationId(source.getId());
 		if(percentageResult.isSuccess()){
-			result.setCompletedPercentage(percentageResult.getSuccessObject().getCompletedPercentage());
+			result.setCompletedPercentage(percentageResult.getSuccessObject().getCompletedPercentage().intValue());
 		}
 		
 		result.setApplicationStatus(source.getApplicationStatus().getId());
